backend/requirements.txt
Flask>=2.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
google-generativeai>=0.3.0
python-multipart>=0.0.6
Werkzeug>=2.3.0
openpyxl>=3.1.0
pdfplumber>=0.9.0
python-magic>=0.4.27
flask-cors>=4.0.0
streamlit>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
nltk>=3.8.1
spacy>=3.6.0
contractions>=0.1.73
inflect>=7.0.0
beautifulsoup4>=4.12.0